requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
apscheduler==3.10.4
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
eventlet==0.35.1
